refactor: organize code into proper folder structure
- Extract components: Navbar, WelcomeScreen, LoadingScreen, QuizSection, ResultsSection
- Create custom hooks: useQuiz, useDarkMode
- Add utility functions for score calculation and progress tracking
- Centralize constants for categories, difficulties, and question options
- Improve code maintainability and readability
- Reduce App.jsx from 507 lines to 99 lines
