+++ src/components/Navbar.jsx
@@ -1,0 +1,45 @@
+export default function Navbar({ isDarkMode, toggleDarkMode, onResetQuiz }) {
+  return (
+    <nav className="floating-nav">
+      <div className="nav-brand">
+        <span className="nav-logo">🧠</span>
+        <span className="nav-title">Inquizzitive</span>
+      </div>
+      <div className="nav-links">
+        <button className="nav-btn" onClick={onResetQuiz}>
+          Practice
+        </button>
+        <button className="nav-btn">Dashboard</button>
+        <button className="nav-btn nav-btn-primary">Sign In</button>
+
+        <button
+          onClick={toggleDarkMode}
+          className="ml-2 p-2 rounded-lg bg-white/10 hover:bg-white/20 backdrop-blur-sm transition-all duration-300 border border-white/20 hover:border-white/30"
+          aria-label="Toggle dark mode"
+        >
+          {isDarkMode ? (
+            <svg
+              className="w-5 h-5 text-yellow-300"
+              fill="currentColor"
+              viewBox="0 0 20 20"
+            >
+              <path
+                fillRule="evenodd"
+                d="M10 2a1 1 0 011 1v1a1 1 0 11-2 0V3a1 1 0 011-1zm4 8a4 4 0 11-8 0 4 4 0 018 0zm-.464 4.95l.707.707a1 1 0 001.414-1.414l-.707-.707a1 1 0 00-1.414 1.414zm2.12-10.607a1 1 0 010 1.414l-.706.707a1 1 0 11-1.414-1.414l.707-.707a1 1 0 011.414 0zM17 11a1 1 0 100-2h-1a1 1 0 100 2h1zm-7 4a1 1 0 011 1v1a1 1 0 11-2 0v-1a1 1 0 011-1zM5.05 6.464A1 1 0 106.465 5.05l-.708-.707a1 1 0 00-1.414 1.414l.707.707zm1.414 8.486l-.707.707a1 1 0 01-1.414-1.414l.707-.707a1 1 0 011.414 1.414zM4 11a1 1 0 100-2H3a1 1 0 000 2h1z"
+                clipRule="evenodd"
+              />
+            </svg>
+          ) : (
+            <svg
+              className="w-5 h-5 text-gray-300"
+              fill="currentColor"
+              viewBox="0 0 20 20"
+            >
+              <path d="M17.293 13.293A8 8 0 016.707 2.707a8.001 8.001 0 1010.586 10.586z" />
+            </svg>
+          )}
+        </button>
+      </div>
+    </nav>
+  );
+}
