+++ src/constants/quiz.jsx
@@ -1,0 +1,18 @@
+export const CATEGORIES = [
+  "Current Affairs",
+  "Geography", 
+  "History",
+  "Indian Defence",
+  "Politics",
+];
+
+export const DIFFICULTIES = ["Easy", "Medium", "Hard"];
+
+export const QUESTION_OPTIONS = [
+  { value: 5, label: "5 Questions (2.5 min)" },
+  { value: 10, label: "10 Questions (5 min)" },
+  { value: 15, label: "15 Questions (7.5 min)" },
+  { value: 20, label: "20 Questions (10 min)" },
+];
+
+export const SECONDS_PER_QUESTION = 30;
